pyproject.toml
[build-system]
requires = ["hatchling>=1.8.0"]
build-backend = "hatchling.build"

[project]
name = "chat2rag"
dynamic = ["version"]
description = "Chat with RAG"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Jacob Zheng", email = "1453366421@qq.com" }]
keywords = ["rag", "llm", "chatbot", "tts", "asr"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # 页面UI
    "streamlit==1.50.0",
    "requests>=2.31.0",
    "streamlit-antd-components>=0.3.2",
    "openpyxl>=3.1.5",
    "streamlit_cookie==0.1.0",

    # 核心RAG组件
    "haystack-ai==2.19.0",
    "qdrant-haystack==9.2.0",
    "mcp-haystack==0.8.0",

    # 监控和性能
    "openinference-instrumentation-haystack==0.1.27",
    "opentelemetry-sdk==1.38.0",
    "opentelemetry-exporter-otlp==1.38.0",
    "arize-phoenix==12.9.0",
    "pyinstrument==5.1.1",
    "plotly==6.3.0",
    "xlrd",

    # 数据库
    "tortoise-orm[asyncpg]==0.25.1",
    "aerich",

    # Web服务
    "fastapi==0.120.3",
    "uvicorn[standard]==0.34.0",

    # 工具库
    "pyhumps==3.8.0",
    "python-dotenv>=1.0.0",
    "fuzzywuzzy==0.18.0",
    "python-Levenshtein>=0.26.1",
    "jsonref==1.1.0",
    "jsonschema>=4.23.0",

    # 农历
    # "zhdate==0.1",
    # "lunar-python==1.3.12",

    # TTS 服务
    # "kokoro==0.7.16"
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "responses==0.25.7",
    "black",
    "isort",
    "mypy",
    "pre-commit",
]
prod = [
    "gunicorn[standard]==23.2.0",
]

[tool.hatch.version]
path = "VERSION.txt"
pattern = "(?P<version>.+)"

[tool.hatch.build.targets.wheel]
packages = ["chat2rag"]

[project.urls]
"GitHub: issues" = "https://github.com/huxiaolongyin/Chat2RAG/issues"
"GitHub: repo" = "https://github.com/huxiaolongyin/Chat2RAG"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
select = ["E", "F", "B"]
ignore = ["E501"]
target-version = "py311"

[tool.aerich]
tortoise_orm = "chat2rag.config.Config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."